requirements.txt
Flask
Flask-Mail
cryptography
python-dotenv
